Created Wait object. Fixed some bugs regarding output messages. Made some slight optimizations and fixed clang warnings. Made preparations for large data example.

+++ framework/areg/base/private/WriteConverter.cpp
@@ -51,43 +51,43 @@ bool WriteConverter::setString( const String & in_value )
 
 bool WriteConverter::setBool( bool in_value )
 {
-    static constexpr unsigned int size = sizeof(bool);
+    static constexpr unsigned int size = static_cast<unsigned int>(sizeof(bool));
     return (mWriteStream.write(reinterpret_cast<const unsigned char *>(&in_value), size) == size);
 }
 
 bool WriteConverter::setChar( char in_value )
 {
-    static constexpr unsigned int size = sizeof(char);
+    static constexpr unsigned int size = static_cast<unsigned int>(sizeof(char));
     return (mWriteStream.write(reinterpret_cast<const unsigned char *>(& in_value), size) == size);
 }
 
 bool WriteConverter::setChar( wchar_t in_value )
 {
-    static constexpr unsigned int size = sizeof(wchar_t);
+    static constexpr unsigned int size = static_cast<unsigned int>(sizeof(wchar_t));
     return (mWriteStream.write(reinterpret_cast<const unsigned char *>(&in_value), size) == size);
 }
 
 bool WriteConverter::setShort( short in_value )
 {
-    static constexpr unsigned int size = sizeof(short);
+    static constexpr unsigned int size = static_cast<unsigned int>(sizeof(short));
     return (mWriteStream.write(reinterpret_cast<const unsigned char *>(&in_value), size) == size);
 }
 
 bool WriteConverter::setInt( int in_value )
 {
-    static constexpr unsigned int size = sizeof(int);
+    static constexpr unsigned int size = static_cast<unsigned int>(sizeof(int));
     return (mWriteStream.write(reinterpret_cast<const unsigned char *>(&in_value), size) == size);
 }
 
 bool WriteConverter::setInt64( int64_t in_value )
 {
-    static constexpr unsigned int size = sizeof(int64_t);
+    static constexpr unsigned int size = static_cast<unsigned int>(sizeof(int64_t));
     return (mWriteStream.write(reinterpret_cast<const unsigned char *>(&in_value), size) == size);
 }
 
 bool WriteConverter::setFloat( float in_value )
 {
-    static constexpr unsigned int size = sizeof(float);
+    static constexpr unsigned int size = static_cast<unsigned int>(sizeof(float));
     return (mWriteStream.write(reinterpret_cast<const unsigned char *>(&in_value), size) == size);
 }
 
@@ -99,24 +99,24 @@ bool WriteConverter::setDouble( double in_value )
 
 bool WriteConverter::setString( const char * in_value )
 {
-    const unsigned int size = in_value != nullptr ? (NEString::getStringLength<char>(in_value) + 1) * sizeof(char) : 0;
+    const unsigned int size = in_value != nullptr ? static_cast<unsigned int>((NEString::getStringLength<char>(in_value) + 1) * sizeof(char)) : 0u;
     return (size != 0 ? mWriteStream.write(reinterpret_cast<const unsigned char *>(in_value), size) == size : false);
 }
 
 bool WriteConverter::setString( const wchar_t * in_value )
 {
-    const unsigned int size = in_value != nullptr ? (NEString::getStringLength<wchar_t>(in_value) + 1) * sizeof(wchar_t) : 0;
+    const unsigned int size = in_value != nullptr ? static_cast<unsigned int>((NEString::getStringLength<wchar_t>(in_value) + 1) * sizeof(wchar_t)) : 0u;
     return (size != 0 ? mWriteStream.write(reinterpret_cast<const unsigned char *>(&in_value), size) == size : false);
 }
 
 bool WriteConverter::appendString( const char * in_value )
 {
-    const unsigned int size = in_value != nullptr ? NEString::getStringLength<char>(in_value) * sizeof(char) : 0;
+    const unsigned int size = in_value != nullptr ? static_cast<unsigned int>(NEString::getStringLength<char>(in_value) * sizeof(char)) : 0;
     return (size != 0 ? mWriteStream.write(reinterpret_cast<const unsigned char *>(in_value), size) == size : false);
 }
 
 bool WriteConverter::appendString( const wchar_t * in_value )
 {
-    const unsigned int size = in_value != nullptr ? NEString::getStringLength<wchar_t>(in_value) * sizeof(wchar_t) : 0;
+    const unsigned int size = in_value != nullptr ? static_cast<unsigned int>(NEString::getStringLength<wchar_t>(in_value) * sizeof(wchar_t)) : 0;
     return (size != 0 ? mWriteStream.write(reinterpret_cast<const unsigned char *>(in_value), size) == size : false);
 }
